refactor(routes): update imports to use plugins instead of features
Update all route files and layout to import from the new plugins directory structure.
