+++ client/src/app/layout.tsx
@@ -2,7 +2,7 @@ import AppBar from '@/components/app-bar/app-bar'
 import Footer from '@/components/footer'
 import QueryClientCustomProvider from '@/config/query-client-custom-provider'
 import TEXT from '@/constants/text'
-import { AuthProvider } from '@/features/auth/index.client'
+import { AuthProvider } from '@/plugins/auth/index.client'
 import fontService from '@/font/font.service'
 import type { Metadata } from 'next'
 import './globals.css'
